implement complete keyboard handling, win32 support ?
in the usual style of mine, fully written but not yet run, this intermittent commit has complete docstrings for the interface in mind.

I haven't a win32 platform, but diligent effort was made to support it.
This commit also includes a possible fix for windows systems to determine the terminal height and width without calls to fcntl.

- remove unused defaultdict import
- refactor input stream key capability from __init__ to
  _init_keystrokes()
- remove redundant input key capability definitions
- store _state_echo and _state_canonical; necessary for win32 systems
- refactor term.stream to term.o_stream to accompany term.i_stream
- refactor documented @property, do_style
- refactor _resolve_multibyte to _resolve_msb, consistent reference
  'MSB' can be either 'escape sequence' for application keys, or utf-8 bytes
  as part of a hamsterface emoticon.