seperate test files for terminal classification
By seperating each "classification" of terminal types, and ceasing to re-instantiate Terminal() within the same process, we avoid issue #33, where subsequent calls to setupterm() do not have any effect with the underline C-land "cur_term" terminal structure; that is, the first (TERMINAL *) cur_term structure initialized is -always- the structure referred to in subsequent capability lookups, regardless to the number of times setupterm() is called, or the values used.

Also included is a fix for terminal height/width, when stdout is piped to stdin of another program, the terminal size is indeterminate. In this situation, we use the LINES and COLUMNS environment values.

Also of note, many previously passing test cases could not pass when using the Terminal() of kind=, stream=, and force_styling= values intended.

One instance of this is the .number_of_colors; which previously _always_ returned 256, even when stream was StringIO and force_styling was non-True. This method was modified to check the value of 'does_styling', and return 0 if not True.