emit warnings for issue #33, and LINES/COLUMNS
previously test cases were passing that probably should not have, due to issue #33; that is, that a value of type 'int' should always be returned; it isn't when the output stream is not stdout and stdout ("init_descriptor") is also not a terminal. This resolves this issue by using the environment values of COLUMNS and LINES, defaulting to 80x24. If the values of COLUMNS and LINES are not integers, a warning is emitted.