enforce setupterm() as a singleton of ``kind``
emit a warning when a Terminal instance is instantiated with a different ``kind`` variable as previously used, explaining that only the terminal capabilities of the first instance will be used. As this cannot be worked around in Python (problem exists in _cursesmodule.c), we simply emit a warning.

Those who *really* need more than one terminal kind are going to have to perform terminal capabilities in subprocesses. Test cases are a great example of that, a pull request doing so will come later.

(Another example is my own esoteric project of a telnet server, serving different capabilities for different clients in different sub-processes each, by using the multiprocess module.)