test NullCallableString more closely
I believe once this is executed in a sub-process, it may actually fail, but I could be wrong.

Regardless, test more variations, and force_styling to None to more guarantee setupterm() has not been called first, though in this branch, it has been anyway and we can't really help that.