Implement as_subprocess decorator
This decorator executes the decorated function within a child process. The child process raises an exception for failed test cases, writing to (shared) stdout/err, which flows through to the controlling process. Otherwise the an exit status of 0 is emitted.

The parent process, on the other hand, waits for the exit signal of the child, and asserts that the exit code is 0, causing a test case to fail if otherwise.