remove "else:" branch, unicode fixes
good ol' fork() -- the child process within the first if: statement will always exit. So there is no need for an 'else' statement for the parent -- any code that continues outside of the 'if:' is the parent.

Also, it presumed the child process is writing to the pseudo terminal using utf-8 encoding, and decoded as such. Its up to basic python & print to try to display whatever is there to the system's locale. For python2.6, this could raise another exception, but python3 generally knows what's up.