implement height and width fallback for non-tty
once the height and width test is resolved to actually test, we
notice that running `nosetests 2>&1 | less' & etc. will fail:
the ioctl for TIOCGWINSZ fails for non-ttys (such as used by travis CI).
So, we fall-through to the LINES, COLUMNS environment variables, with
default values of (24, 80). (24, 80) has been a fairly standard
screensize for IBM PC-DOS and Apple ][ (beginning with the 80-column
character card), and C128. It is also the default for xterm, and many
classic terminals (such as a vt220) or emulating terminals (such as
telix), or bulletin board servers (such as teleguard) where a 'status
line' is also present.  In reality, these screens are capable of *25*
lines, but the 25th line is reserved for the status line.

Irregardless, non-zero must be returned for 'height' and 'width'
properties, as a value of 0 may become a "divide by zero" error
for mathematical operations that make use of the terminal height or
width in scripts that the user may chose to pipe to 'less -r' or some
such. In these situations, even though a value is returned, operations
such as 'move(x, y)' would still become 'stripped' due to 'is_a_tty'
becoming False, so there is no actual harm in providing a terminal size
that is not legal.