encode tparm output as latin1
as explained in the comments above the decode().encode() wrapper,
certain terminal kinds, such as 'avatar' or 'kermit' emit 8-bit bytes
that are not legal utf-8 tender. this issue is resolved by encoding
these values as latin1, which leaves their values unmolested. This is
exaplified by a gist: https://gist.github.com/jquast/5649654

If you don't believe me, try kind='kermit' or kind='avatar', along with
t.cup(n, n). Unfortunately, the state of the test cases in the master
branch (issue #33) does not allow to add a test case in this branch
until that one is pulled to master.

Hell, I'll go ahead and make a third (and final) pull request for that one,