Closes issue #6
1. make ParameterizingString descriptive and kind

esp. regarding multi-arguments, change "if len == 1"
test to be "if len", so that we can more kindly
report to the user about their failed attempt
at blue("1", "2") whoops.

2. compoundables and derivatives got out of order
when we refactored these into a submodule, the
red_on_black wasn't matching, because the on_black
derivatives weren't calculated in the correct order.