Backmerge rejected upstream enhancements
Version 1.9.6 readyness: A branch in the upstream
repository to merge all of our work has stalled.

https://github.com/erikrose/blessings/pull/104

This squashed commit returns all such work into the
mainline branch of 'blessed', so that work may progress
which has otherwise stagnated since January of this year.

Once these commits are squashed and boiled out, most of
which are renames that are later reversed, very little
remains that warrant a release.  We are left with
the following changes:

- Bugfix? Remove 'unknown' terminal type.
- Add dependency 'six' for 'string_type'.
- Improvements in documentation.
- Improvements in test and build chain.