Bring Python versions under test up to date.
To heck with 3.2. Keep 3.3 just in case there's something special about it; I don't remember. Add 3.6 because it's the newest.

Move installation stuff to a Travis `install` block so it doesn't show up as a test failure if it breaks. Stop using pip's `--use-mirrors` option, which is gone in modern pips.